requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
plotly>=5.18.0

# Time series analysis
statsmodels>=0.14.0
prophet>=1.1.5

# Machine learning
scikit-learn>=1.3.0
torch>=2.0.0

# Time series specialized
tslearn>=0.6.2
pyts>=0.13.0
stumpy>=1.13.0

# Real data sources
yfinance>=0.2.0

# LLM integration
openai>=1.3.0

# Environment
python-dotenv>=1.0.0
